tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-playwright==0.4.3
playwright==1.40.0
httpx==0.25.2
python-docx==1.0.1
docx2pdf==0.1.8
sqlalchemy==2.0.23
pydantic==2.5.2
fastapi==0.104.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 